Remove directory size check
Each music directory was previously checked and only processed when its
total size was above 10 MB. The purpose of this was to ignore empty
ghost directories that belonged to deleted torrents. Considering that
WhatManager currently removes all files without a trace, this is not
necessary. Also, there can exist legitimate torrents with sizes below 10
MB. Another reason to remove this code is that size calculations can
take a lot of time to complete.